pipes/gphotos/pyproject.toml
[project]
name = "shenas-pipe-gphotos"
dynamic = ["version"]
description = "Google Photos data connector for shenas"
requires-python = ">=3.11"
dependencies = [
    "dlt[duckdb]>=1.24.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "pendulum>=3.0.0",
    "shenas-pipe-core",
]

[project.entry-points."shenas.pipes"]
gphotos = "shenas_pipes.gphotos.cli:app"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
tag-pattern = "pipe-gphotos/v(?P<version>[^+]+)"
raw-options.root = "../.."
raw-options.git_describe_command = ["git", "describe", "--dirty", "--tags", "--long", "--match", "pipe-gphotos/v*"]

[tool.hatch.build.targets.wheel]
packages = ["src/shenas_pipes"]
